Copying wwise audio assets from external svn.

+++ project/iosCopyResources.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 set -e
+set -x
+
 SRCDIR="$1"
 ASSETSRDIR="$2"
 DSTDIR="$3"
@@ -18,6 +20,7 @@ TOPLEVEL=`$GIT rev-parse --show-toplevel`
 cd $TOPLEVEL
 
 SPHINXDIR="$TOPLEVEL"/generated/resources/pocketsphinx
+ASSETS_SOUND_DIR="$TOPLEVEL"/EXTERNALS/cozmosoundbanks/GeneratedSoundBanks/iOS
 DSTDIR_META=$DSTDIR/cozmo_resources/config/
 DSTDIR_SPHINX=$DSTDIR/cozmo_resources/pocketsphinx/
 DSTDIR_ASSET=$DSTDIR/cozmo_resources/assets/
@@ -48,6 +51,11 @@ if [ -z "DSTDIR" ] ; then
 	exit 1;
 fi
 
+if [ ! -d "$ASSETS_SOUND_DIR" ] ; then
+	echo "wwise sound assets not found $ASSETS_SOUND_DIR"
+	exit 1;
+fi
+
 # copy resources
 mkdir -p "$DSTDIR_META"
 rsync -r -t --exclude=".*" --delete $SRCDIR/ $DSTDIR_META/
@@ -59,6 +67,9 @@ if [ -d "$ASSETSRDIR" ] ; then
   rsync -r -t --exclude=".*" --delete $ASSETSRDIR/ $DSTDIR_ASSET/
 fi
 
+if [ -d "$ASSETS_SOUND_DIR" ]; then
+  rsync -v -r -t --exclude=".*" $ASSETS_SOUND_DIR/ $DSTDIR_ASSET/sounds
+fi
 
 echo asset: $DSTDIR_ASSET
 echo meta: $DSTDIR_META
